Refactor fileeditor-extension for modularization
An initial POC for modularizing the features added in the activate function
to allow other extensions to call them. This inital PR sets the ground work
and includes the modularization of the `menu.addItems` code block.

Part of the ongoing Issue #6901
