+++ packages/fileeditor-extension/package.json
@@ -47,6 +47,7 @@
     "@jupyterlab/launcher": "^1.0.1",
     "@jupyterlab/mainmenu": "^1.0.1",
     "@jupyterlab/statusbar": "^1.0.1",
+    "@phosphor/commands": "^1.6.3",
     "@phosphor/coreutils": "^1.3.1",
     "@phosphor/widgets": "^1.8.0"
   },
